+++ app/src/main/java/com/alphawallet/app/service/AssetDefinitionService.java
@@ -415,7 +415,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return td;
     }
 
-    private String getTSDataKey(int chainId, String address)
+    private String getTSDataKey(long chainId, String address)
     {
         return address + "-" + chainId;
     }
@@ -555,7 +555,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
     }
 
     @Override
-    public long getLastTokenUpdate(int chainId, String address)
+    public long getLastTokenUpdate(long chainId, String address)
     {
         long txUpdateTime = 0;
         Token token = tokensService.getToken(chainId, address);
@@ -572,7 +572,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
     {
         String addr = null;
         TokenDefinition td = null;
-        int chainId = origin.addresses.keySet().iterator().next();
+        long chainId = origin.addresses.keySet().iterator().next();
         if (origin.addresses.get(chainId).size() > 0) addr = origin.addresses.get(chainId).get(0);
         if (addr != null) td = getAssetDefinition(chainId, addr);
         if (td != null)
@@ -686,7 +686,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
                 == PackageManager.PERMISSION_GRANTED;
     }
 
-    private TokenDefinition getDefinition(int chainId, String address)
+    private TokenDefinition getDefinition(long chainId, String address)
     {
         if (address.equalsIgnoreCase(tokensService.getCurrentAddress())) address = "ethereum";
         TokenDefinition result = null;
@@ -732,7 +732,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return result;
     }
 
-    public TokenScriptFile getTokenScriptFile(int chainId, String address)
+    public TokenScriptFile getTokenScriptFile(long chainId, String address)
     {
         //pull from database
         if (address.equalsIgnoreCase(tokensService.getCurrentAddress())) address = "ethereum";
@@ -757,7 +757,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
      * @param address
      * @return
      */
-    public TokenDefinition getAssetDefinition(int chainId, String address)
+    public TokenDefinition getAssetDefinition(long chainId, String address)
     {
         TokenDefinition assetDef = null;
         if (address == null) return null;
@@ -777,7 +777,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return assetDef; // if nothing found use default
     }
 
-    public Single<TokenDefinition> getAssetDefinitionASync(int chainId, final String address)
+    public Single<TokenDefinition> getAssetDefinitionASync(long chainId, final String address)
     {
         if (address == null) return Single.fromCallable(TokenDefinition::new);
         String contractName = address;
@@ -813,7 +813,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         }
     }
 
-    public String getTokenName(int chainId, String address, int count)
+    public String getTokenName(long chainId, String address, int count)
     {
         String tokenName = null;
         if (address.equalsIgnoreCase(tokensService.getCurrentAddress())) address = "ethereum";
@@ -832,7 +832,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return tokenName;
     }
 
-    public Token getTokenFromService(int chainId, String address)
+    public Token getTokenFromService(long chainId, String address)
     {
         return tokensService.getToken(chainId, address);
     }
@@ -846,7 +846,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
      */
     public String getIssuerName(Token token)
     {
-        int chainId = token.tokenInfo.chainId;
+        long chainId = token.tokenInfo.chainId;
         String address = token.tokenInfo.address;
 
         String issuer = token.getNetworkName();
@@ -1091,7 +1091,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
             if (holdingContracts != null)
             {
                 //some Android versions don't have stream()
-                for (int network : holdingContracts.addresses.keySet())
+                for (long network : holdingContracts.addresses.keySet())
                 {
                     for (String address : holdingContracts.addresses.get(network))
                     {
@@ -1129,7 +1129,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return td;
     }
 
-    private void updateRealmForBundledScript(int chainId, String address, String asset, TokenDefinition td)
+    private void updateRealmForBundledScript(long chainId, String address, String asset, TokenDefinition td)
     {
         realmManager.getRealmInstance(ASSET_DEFINITION_DB).executeTransactionAsync(r -> {
             String entryKey = getTSDataKey(chainId, address);
@@ -1284,7 +1284,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
 
     private EthFilter getEventFilter(EventDefinition ev) throws Exception
     {
-        int chainId = ev.getEventChainId();
+        long chainId = ev.getEventChainId();
         String address = ev.getEventContractAddress();
 
         Token originToken = tokensService.getToken(chainId, address);
@@ -1297,7 +1297,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
     private String processLogs(EventDefinition ev, List<EthLog.LogResult> logs, String walletAddress)
     {
         if (logs.size() == 0) return ""; //early return
-        int chainId = ev.contract.addresses.keySet().iterator().next();
+        long chainId = ev.contract.addresses.keySet().iterator().next();
         Web3j web3j = getWeb3jService(chainId);
 
         String firstTxHash = "";
@@ -1352,7 +1352,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         ev.readBlock = readBlock.add(BigInteger.ONE);
         try (Realm realm = realmManager.getRealmInstance(walletAddress))
         {
-            int chainId = ev.getEventChainId();
+            long chainId = ev.getEventChainId();
             String eventAddress = ev.getEventContractAddress();
             String eventName = ev.activityName != null ? ev.activityName : ev.attributeName;
             String databaseKey = TokensRealmSource.eventBlockKey(chainId, eventAddress, ev.type.name, ev.filter);
@@ -1617,7 +1617,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return file;
     }
 
-    public boolean hasDefinition(int chainId, String address)
+    public boolean hasDefinition(long chainId, String address)
     {
         boolean hasDefinition = false;
         if (address.equalsIgnoreCase(tokensService.getCurrentAddress())) address = "ethereum";
@@ -1639,7 +1639,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         assetChecked.clear();
     }
 
-    public boolean hasTokenView(int chainId, String address, String type)
+    public boolean hasTokenView(long chainId, String address, String type)
     {
         if (address.equalsIgnoreCase(tokensService.getCurrentAddress())) address = "ethereum";
         try (Realm realm = realmManager.getRealmInstance(ASSET_DEFINITION_DB))
@@ -1652,7 +1652,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         }
     }
 
-    public String getTokenView(int chainId, String contractAddr, String type)
+    public String getTokenView(long chainId, String contractAddr, String type)
     {
         String viewHTML = "";
         TokenDefinition td = getAssetDefinition(chainId, contractAddr);
@@ -1664,7 +1664,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return viewHTML;
     }
 
-    public String getTokenViewStyle(int chainId, String contractAddr, String type)
+    public String getTokenViewStyle(long chainId, String contractAddr, String type)
     {
         String styleData = "";
         TokenDefinition td = getAssetDefinition(chainId, contractAddr);
@@ -1676,7 +1676,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return styleData;
     }
 
-    public List<Attribute> getTokenViewLocalAttributes(int chainId, String contractAddr)
+    public List<Attribute> getTokenViewLocalAttributes(long chainId, String contractAddr)
     {
         TokenDefinition td = getAssetDefinition(chainId, contractAddr);
         List<Attribute> results = new ArrayList<>();
@@ -1689,7 +1689,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return results;
     }
 
-    public Map<String, TSAction> getTokenFunctionMap(int chainId, String contractAddr)
+    public Map<String, TSAction> getTokenFunctionMap(long chainId, String contractAddr)
     {
         TokenDefinition td = getAssetDefinition(chainId, contractAddr);
         if (td != null)
@@ -1942,7 +1942,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return observer;
     }
 
-    public Single<XMLDsigDescriptor> getSignatureData(int chainId, String contractAddress)
+    public Single<XMLDsigDescriptor> getSignatureData(long chainId, String contractAddress)
     {
         return Single.fromCallable(() -> {
             XMLDsigDescriptor sigDescriptor = new XMLDsigDescriptor();
@@ -2152,7 +2152,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         String[] contractDetails = eventData.getInstanceKey().split("-");
         if (contractDetails.length != 5) return;
         String eventAddress = contractDetails[0];
-        int chainId = Integer.parseInt(contractDetails[1]);
+        long chainId = Long.parseLong(contractDetails[1]);
         String eventId = eventData.getFunctionId();
 
         String eventKey = EventDefinition.getEventKey(chainId, eventAddress, eventId, null);
@@ -2242,7 +2242,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
      * Get all the magic values - eg native crypto balances for all chains
      * @return
      */
-    public String getMagicValuesForInjection(int chainId) throws Exception
+    public String getMagicValuesForInjection(long chainId) throws Exception
     {
         String walletBalance = "walletBalance";
         String prefix = "web3.eth";
@@ -2447,7 +2447,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         });
     }
 
-    public Single<TokenDefinition> checkServerForScript(int chainId, String address)
+    public Single<TokenDefinition> checkServerForScript(long chainId, String address)
     {
         TokenScriptFile tf = getTokenScriptFile(chainId, address);
         if (tf != null && !isInSecureZone(tf)) return Single.fromCallable(TokenDefinition::new); //early return for debug script check
@@ -2460,7 +2460,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
                 .observeOn(AndroidSchedulers.mainThread());
     }
 
-    public void storeTokenViewHeight(int chainId, String address, int listViewHeight)
+    public void storeTokenViewHeight(long chainId, String address, int listViewHeight)
     {
         try (Realm realm = realmManager.getRealmInstance(tokensService.getCurrentAddress()))
         {
@@ -2486,7 +2486,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         }
     }
 
-    public String getTokenImageUrl(int networkId, String address)
+    public String getTokenImageUrl(long networkId, String address)
     {
         String url = "";
         String instanceKey = address.toLowerCase() + "-" + networkId;
@@ -2522,7 +2522,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         return tURL;
     }
 
-    public void storeImageUrl(int chainId, String imageUrl)
+    public void storeImageUrl(long chainId, String imageUrl)
     {
         String tokenAddress = Utils.getTokenAddrFromAWUrl(imageUrl);
         if (!TextUtils.isEmpty(tokenAddress))
@@ -2531,7 +2531,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         }
     }
 
-    public Single<Integer> fetchViewHeight(int chainId, String address)
+    public Single<Integer> fetchViewHeight(long chainId, String address)
     {
         return Single.fromCallable(() -> {
             try (Realm realm = realmManager.getRealmInstance(tokensService.getCurrentAddress()))
@@ -2567,7 +2567,7 @@ public class AssetDefinitionService implements ParseResult, AttributeInterface
         });
     }
 
-    private String tokenSizeDBKey(int chainId, String address)
+    private String tokenSizeDBKey(long chainId, String address)
     {
         return "szkey-" + chainId + "-" + address.toLowerCase();
     }
